requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cloudinary==1.28.0
colorama==0.4.4
cryptography==36.0.0
Django==3.2.9
django-binary-database-files==1.0.15
django-cloudinary-storage==0.3.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
Flask==2.0.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.1
pycparser==2.21
python-magic==0.4.24
pytz==2021.3
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
